First release of bent plume model
Major model refactoring and first release of the new bent plume model
module, which simulates a Lagrangian plume model for multiphase plumes
in crossflow.  The ./bin examples and ./test unit test are not yet
released for the bent plume model, but the documentation demonstrates
how to use the new codes.