Bent plume model temperature fix
Updated the bent plume model so that the plume temperature is used in
the particle.update() function:  that is, particles are surrounded by
plume fluid, not ambient fluid.  This will mostly allow the model to
turn off heat transfer at the right time.  It also gives a better
estimate of particle properties, but probably not a major change there.